Move the compilation of PCN and SAN regexp out of the Position.ParseMove
To increase speed it's better to compile a regexp once on app start and
then use the compiled output structure.

The underlying structure of ParseMove and originOfPiece is no longer
referenced as pointer. This indicates, that the underlying structure is
not changed when calling ParseMove or originOfPiece.

Unneeded TODO comments in ParseMove description removed. They are fixed
now.