Parse move tests & benchmarks (#19)
* Position.ParseMove tests and benchmarks

To test the Position.ParseMove function, many tests were added to the
new TestParseMove test function. The test cases are in the
testParseMoveGroups variable and are separated by different board
positions on which more moves are tested.

The TestParseLongCastle and TestParseNoPromo test functions were
removed, cause their test counterparts are in the new TestParseMove test
function.

Some unexported test structures and functions were added to make code
readability better.

Benchmarks utilizing the ParseMove tests data were added.

* Fix SAN regexp in Position.ParseMove function

* Fix SAN regexp in Position.ParseMove function

* Promote to Queen if promotion is missing

If a promotion is missing in PCN or SAN input to Position.ParseMove
function, return a queen promotion.

* Fix Position.originOfPiece when more squares are matched as source square

Do not return first matching source square if position is ambiguous and
more origin squares are possible for SAN. Return error instead.

* Move the compilation of PCN and SAN regexp out of the Position.ParseMove

To increase speed it's better to compile a regexp once on app start and
then use the compiled output structure.

The underlying structure of ParseMove and originOfPiece is no longer
referenced as pointer. This indicates, that the underlying structure is
not changed when calling ParseMove or originOfPiece.

Unneeded TODO comments in ParseMove description removed. They are fixed
now.